Move to pytest and specify test_dependencies
