+++ test_dependencies.txt
@@ -1,0 +1,4 @@
+https://github.com/modflowpy/pymake/zipball/master
+https://github.com/modflowpy/flopy/zipball/develop
+pytest
+pytest-cov
